siemens/s97801: address review round 2
- machine.lua: move S97801 back into alphabetical order (it was splitting
  the SA11xx series)
- attach the keyboard layout to the s97801 device (like ie15) so it is
  available when the terminal is used as an rs232 peripheral as well as
  standalone; move the layout to emu/layout and make the tags device-relative
- device lifecycle: initialise m_kbd_rxd in device_resolve_objects and push
  the initial txd state in device_start rather than driving them from
  device_reset; do the rs232 terminal card initial RxD/modem outputs in
  device_start too, to avoid glitches on soft reset
- use explicitly sized integer types for saved state (m_kbd_rxd, m_rxd)
- visually separate configuration from live-signal member functions
- use KEYCODE_00_PAD/KEYCODE_000_PAD/KEYCODE_EQUALS_PAD for the keypad keys

+++ src/devices/machine/s97801_kbd.h
@@ -17,7 +17,10 @@ class s97801_kbd_device : public device_t
 public:
 	s97801_kbd_device(const machine_config &mconfig, const char *tag, device_t *owner, u32 clock = 0);
 
+	// configuration
 	auto txd_handler() { return m_txd_cb.bind(); } // keyboard -> terminal
+
+	// live signals
 	void rxd_w(int state);                         // terminal -> keyboard
 
 protected:
@@ -29,6 +32,7 @@ protected:
 private:
 	void prg_map(address_map &map) ATTR_COLD;
 
+	// live signals
 	u8 bus_r();
 	void bus_w(u8 data);
 	void p1_w(u8 data);
@@ -49,7 +53,7 @@ private:
 	emu_timer *m_bell_timer;
 	u8 m_p1;
 	u8 m_p2;
-	int m_rxd;
+	u8 m_rxd;
 };
 
 DECLARE_DEVICE_TYPE(SIEMENS_97801_KBD, s97801_kbd_device)
